# apps/backend/src/nest/translation/prompt/dto/response/delete-prompt-preset.response.dto.ts
import { BaseResponseDto } from '@apps/common/dist/dto/base-response.dto';

export class DeletePromptPresetResponseDto extends BaseResponseDto {
  // 삭제 성공/실패 여부와 메시지만 포함
}