# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/prompt-presets": {
      "get": {
        "operationId": "PromptPresetController_getPromptPresets",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "프롬프트 프리셋 타입 필터",
            "schema": {
              "example": "text",
              "type": "string",
              "enum": [
                "text",
                "image"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "프롬프트 프리셋 목록을 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPromptPresetsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "prompt-presets"
        ]
      },
      "post": {
        "operationId": "PromptPresetController_createPromptPreset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptPresetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "프롬프트 프리셋을 생성합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptPresetResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "prompt-presets"
        ]
      }
    },
    "/prompt-presets/{id}": {
      "get": {
        "operationId": "PromptPresetController_getPromptPresetDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "프롬프트 프리셋 상세 정보를 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPromptPresetDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "prompt-presets"
        ]
      },
      "patch": {
        "operationId": "PromptPresetController_updatePromptPreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptPresetBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "프롬프트 프리셋을 업데이트합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePromptPresetResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "prompt-presets"
        ]
      },
      "delete": {
        "operationId": "PromptPresetController_deletePromptPreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "프롬프트 프리셋을 삭제합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePromptPresetResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "prompt-presets"
        ]
      }
    },
...
  },
  "info": {
    "title": "FormattedDocs API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PromptPresetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "프롬프트 프리셋 ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "프롬프트 프리셋 이름",
            "example": "기본 프롬프트"
          },
          "type": {
            "type": "string",
            "description": "프롬프트 프리셋 타입",
            "enum": [
              "text",
              "image"
            ],
            "example": "text"
          }
        },
        "required": [
          "id",
          "name",
          "type"
        ]
      },
      "GetPromptPresetsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "presets": {
            "description": "프롬프트 프리셋 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PromptPresetDto"
            }
          }
        },
        "required": [
          "success",
          "presets"
        ]
      },
      "PromptPresetDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "프롬프트 프리셋 ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "프롬프트 프리셋 이름",
            "example": "기본 프롬프트"
          },
          "prompt": {
            "type": "string",
            "description": "프롬프트 내용",
            "example": "당신은 번역 전문가입니다."
          },
          "type": {
            "type": "string",
            "description": "프롬프트 프리셋 타입",
            "enum": [
              "text",
              "image"
            ],
            "example": "text"
          }
        },
        "required": [
          "id",
          "name",
          "prompt",
          "type"
        ]
      },
      "GetPromptPresetDetailResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "preset": {
            "description": "프롬프트 프리셋 상세 정보",
            "allOf": [
              {
                "$ref": "#/components/schemas/PromptPresetDetailDto"
              }
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "CreatePromptPresetRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "프롬프트 프리셋 이름",
            "example": "기본 프롬프트"
          },
          "prompt": {
            "type": "string",
            "description": "프롬프트 내용",
            "example": "당신은 번역 전문가입니다."
          },
          "type": {
            "type": "string",
            "description": "프롬프트 프리셋 타입",
            "enum": [
              "text",
              "image"
            ],
            "example": "text"
          }
        },
        "required": [
          "name",
          "prompt",
          "type"
        ]
      },
      "CreatePromptPresetResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "preset": {
            "description": "생성된 프롬프트 프리셋",
            "allOf": [
              {
                "$ref": "#/components/schemas/PromptPresetDto"
              }
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "UpdatePromptPresetBodyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "프롬프트 프리셋 이름",
            "example": "수정된 프롬프트"
          },
          "prompt": {
            "type": "string",
            "description": "프롬프트 내용",
            "example": "당신은 번역 전문가입니다."
          },
          "type": {
            "type": "string",
            "description": "프롬프트 프리셋 타입",
            "enum": [
              "text",
              "image"
            ],
            "example": "text"
          }
        }
      },
      "UpdatePromptPresetResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "preset": {
            "description": "업데이트된 프롬프트 프리셋",
            "allOf": [
              {
                "$ref": "#/components/schemas/PromptPresetDto"
              }
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "DeletePromptPresetResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}